requirements.txt
django==3.2.5
djangorestframework
django-cors-headers
django-storages
django-rest-auth
python-memcached
img2pdf
python-decouple